src/db_scripts/output_sdk/ai_defense/pyproject.toml
[tool.poetry]
name = "ai-defense-ap-is-client"
version = "1.0.0.2025.02.28"
description = "A client library for accessing AI Defense APIs"
authors = []
readme = "README.md"
packages = [
    {include = "ai_defense_ap_is_client"},
]
include = ["CHANGELOG.md", "ai_defense_ap_is_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
